adding dry run functionality
